classifier_task/requirements.txt
numpy==1.23.5
scikit-learn==1.2.0
pytorch==2.0.0
torchvision==0.15.0
